+++ Headers/AppKit/NSGraphicsContext.h
@@ -279,8 +279,8 @@ APPKIT_EXPORT NSGraphicsContext	*GSCurrentContext(void);
 
 - (void) GSSetFillColorspace: (void *)spaceref;
 - (void) GSSetStrokeColorspace: (void *)spaceref;
-- (void) GSSetFillColor: (const float *)values;
-- (void) GSSetStrokeColor: (const float *)values;
+- (void) GSSetFillColor: (const CGFloat *)values;
+- (void) GSSetStrokeColor: (const CGFloat *)values;
 
 /* ----------------------------------------------------------------------- */
 /* Text operations */
